up: [uiux][arc] sh 実行を /cmd/main.go (package main) に移行, go 実行を /generator.go (elasticsearch2go) ^^ // [uiux][arc] sh 実行, go 実行のどちらでもより使いやすく ^^
